+++ src/Service/Bulk/BulkCreateUsersAssignmentsService.php
@@ -24,13 +24,11 @@ namespace OAT\SimpleRoster\Service\Bulk;
 
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\ORM\EntityNotFoundException;
-use Doctrine\ORM\NonUniqueResultException;
 use OAT\SimpleRoster\Bulk\Operation\BulkOperation;
 use OAT\SimpleRoster\Bulk\Operation\BulkOperationCollection;
 use OAT\SimpleRoster\Bulk\Processor\BulkOperationCollectionProcessorInterface;
 use OAT\SimpleRoster\Bulk\Result\BulkResult;
 use OAT\SimpleRoster\Entity\Assignment;
-use OAT\SimpleRoster\Entity\User;
 use OAT\SimpleRoster\Model\UsernameCollection;
 use OAT\SimpleRoster\Repository\UserRepository;
 use OAT\SimpleRoster\Service\Cache\UserCacheWarmerService;
@@ -41,6 +39,9 @@ class BulkCreateUsersAssignmentsService implements BulkOperationCollectionProces
 {
     /** @var EntityManagerInterface */
     private $entityManager;
+
+    /** @var UserRepository */
+    private $userRepository;
 
     /** @var UserCacheWarmerService */
     private $userCacheWarmerService;
@@ -53,10 +54,12 @@ class BulkCreateUsersAssignmentsService implements BulkOperationCollectionProces
 
     public function __construct(
         EntityManagerInterface $entityManager,
+        UserRepository $userRepository,
         UserCacheWarmerService $userCacheWarmerService,
         LoggerInterface $logger
     ) {
         $this->entityManager = $entityManager;
+        $this->userRepository = $userRepository;
         $this->userCacheWarmerService = $userCacheWarmerService;
         $this->logger = $logger;
     }
@@ -90,13 +93,10 @@ class BulkCreateUsersAssignmentsService implements BulkOperationCollectionProces
 
     /**
      * @throws EntityNotFoundException
-     * @throws NonUniqueResultException
      */
     private function processOperation(BulkOperation $operation, BulkResult $result): void
     {
-        /** @var UserRepository $userRepository */
-        $userRepository = $this->entityManager->getRepository(User::class);
-        $user = $userRepository->findByUsernameWithAssignments($operation->getIdentifier());
+        $user = $this->userRepository->findByUsernameWithAssignments($operation->getIdentifier());
 
         $lastAssignment = $user->getLastAssignment();