+++ tests/Functional/Command/ModifyEntity/LineItem/LineItemChangeStateCommandTest.php
@@ -28,6 +28,7 @@ use Exception;
 use LogicException;
 use Monolog\Logger;
 use OAT\SimpleRoster\Command\ModifyEntity\LineItem\LineItemChangeStateCommand;
+use OAT\SimpleRoster\Entity\LineItem;
 use OAT\SimpleRoster\Generator\LineItemCacheIdGenerator;
 use OAT\SimpleRoster\Repository\LineItemRepository;
 use OAT\SimpleRoster\Tests\Traits\CommandDisplayNormalizerTrait;
@@ -85,20 +86,21 @@ class LineItemChangeStateCommandTest extends KernelTestCase
     /**
      * @dataProvider provideValidArguments
      */
-    public function testItDeactivatesLineItem(array $lineItemIds, string $queryField, array $queryValue): void
+    public function testItDisablesLineItem(array $lineItemIds, string $queryField, array $queryValue): void
     {
         foreach ($lineItemIds as $lineItemId) {
             $lineItem = $this->lineItemRepository->findOneById($lineItemId);
-            self::assertTrue($lineItem->isActive());
+            self::assertTrue($lineItem->isEnabled());
 
             $lineItemCache = $this->resultCache->fetch($this->lineItemCacheIdGenerator->generate($lineItemId));
             $cache = current(current($lineItemCache));
-            self::assertSame('1', $cache['is_active_3']);
+
+            self::assertSame('enabled', $cache['status_3']);
         }
 
         $commandResult = $this->commandTester->execute(
             [
-                'toggle' => 'deactivate',
+                'toggle' => 'disable',
                 'query-field' => $queryField,
                 'query-value' => $queryValue,
             ],
@@ -106,30 +108,30 @@ class LineItemChangeStateCommandTest extends KernelTestCase
         );
         self::assertEquals(0, $commandResult);
 
-        $this->assertMessageDisplays($lineItemIds, 'deactivate');
+        $this->assertMessageDisplays($lineItemIds, 'disable');
 
-        $this->assertLineItems($lineItemIds, 'deactivate', false);
+        $this->assertLineItems($lineItemIds, 'disable', LineItem::STATUS_DISABLED);
     }
 
     /**
      * @dataProvider provideValidArguments
      */
-    public function testItActivatesLineItem(array $lineItemIds, string $queryField, array $queryValue): void
+    public function testItEnablesLineItem(array $lineItemIds, string $queryField, array $queryValue): void
     {
         $this->commandTester->execute(
             [
-                'toggle' => 'deactivate',
+                'toggle' => 'disable',
                 'query-field' => $queryField,
                 'query-value' => $queryValue,
             ],
             ['capture_stderr_separately' => true]
         );
 
-        $this->assertLineItems($lineItemIds, 'deactivate', false);
+        $this->assertLineItems($lineItemIds, 'disable', LineItem::STATUS_DISABLED);
 
         $commandResult = $this->commandTester->execute(
             [
-                'toggle' => 'activate',
+                'toggle' => 'enable',
                 'query-field' => $queryField,
                 'query-value' => $queryValue,
             ],
@@ -137,29 +139,29 @@ class LineItemChangeStateCommandTest extends KernelTestCase
         );
         self::assertEquals(0, $commandResult);
 
-        $this->assertMessageDisplays($lineItemIds, 'activate');
+        $this->assertMessageDisplays($lineItemIds, 'enable');
 
-        $this->assertLineItems($lineItemIds, 'activate', true);
+        $this->assertLineItems($lineItemIds, 'enable', LineItem::STATUS_ENABLED);
     }
 
-    public function testEnsureOnlyOneLineIsDeactivated(): void
+    public function testEnsureOnlyOneLineIsDisabled(): void
     {
         $secondLineItem = $this->lineItemRepository->findOneById(new UuidV6('00000002-0000-6000-0000-000000000000'));
-        self::assertTrue($secondLineItem->isActive());
+        self::assertTrue($secondLineItem->isEnabled());
 
         $this->commandTester->execute(
             [
-                'toggle' => 'deactivate',
+                'toggle' => 'disable',
                 'query-field' => 'id',
                 'query-value' => '00000001-0000-6000-0000-000000000000',
             ],
             ['capture_stderr_separately' => true]
         );
         $firstLineItem = $this->lineItemRepository->findOneById(new UuidV6('00000001-0000-6000-0000-000000000000'));
-        self::assertFalse($firstLineItem->isActive());
+        self::assertFalse($firstLineItem->isEnabled());
 
         $secondLineItem = $this->lineItemRepository->findOneById(new UuidV6('00000002-0000-6000-0000-000000000000'));
-        self::assertTrue($secondLineItem->isActive());
+        self::assertTrue($secondLineItem->isEnabled());
     }
 
     /**
@@ -196,7 +198,7 @@ class LineItemChangeStateCommandTest extends KernelTestCase
         $commandTester = new CommandTester($application->find(LineItemChangeStateCommand::NAME));
 
         $commandTester->execute(
-            ['toggle' => 'deactivate', 'query-field' => 'id', 'query-value' => '00000001-0000-6000-0000-000000000000'],
+            ['toggle' => 'disable', 'query-field' => 'id', 'query-value' => '00000001-0000-6000-0000-000000000000'],
             ['capture_stderr_separately' => true]
         );
 
@@ -261,11 +263,11 @@ class LineItemChangeStateCommandTest extends KernelTestCase
                     'query-field' => 'id',
                     'query-value' => '1',
                 ],
-                'expectedMessage' => 'Invalid toggle argument. Please use: activate, deactivate',
+                'expectedMessage' => 'Invalid toggle argument. Please use: enable, disable',
             ],
             'invalidQueryField' => [
                 'arguments' => [
-                    'toggle' => 'activate',
+                    'toggle' => 'enable',
                     'query-field' => 'invalid-query-field',
                     'query-value' => '1',
                 ],
@@ -277,13 +279,13 @@ class LineItemChangeStateCommandTest extends KernelTestCase
             ],
             'only-toggle' => [
                 'arguments' => [
-                    'toggle' => 'activate',
+                    'toggle' => 'enable',
                 ],
                 'expectedMessage' => 'Not enough arguments (missing: "query-field, query-value")',
             ],
             'with-no-query-value' => [
                 'arguments' => [
-                    'toggle' => 'activate',
+                    'toggle' => 'enable',
                     'query-field' => 'uri',
                 ],
                 'expectedMessage' => 'Not enough arguments (missing: "query-value")',
@@ -294,7 +296,7 @@ class LineItemChangeStateCommandTest extends KernelTestCase
     private function assertMessageDisplays(array $lineItemIds, string $toggle): void
     {
         self::assertStringContainsString(
-            'Simple Roster - Line Item Activator',
+            'Simple Roster - Line item status updater',
             $this->commandTester->getDisplay()
         );
         self::assertStringContainsString(
@@ -307,7 +309,7 @@ class LineItemChangeStateCommandTest extends KernelTestCase
         );
     }
 
-    private function assertLineItems(array $lineItemIds, string $toggle, bool $isActive): void
+    private function assertLineItems(array $lineItemIds, string $toggle, string $expectedLineItemStatus): void
     {
         foreach ($lineItemIds as $lineItemId) {
             $this->assertHasLogRecord(
@@ -323,12 +325,12 @@ class LineItemChangeStateCommandTest extends KernelTestCase
 
             $lineItem = $this->lineItemRepository->findOneById($lineItemId);
 
-            self::assertSame($isActive, $lineItem->isActive());
+            self::assertSame($expectedLineItemStatus, $lineItem->getStatus());
 
             $cache = current(
                 current($this->resultCache->fetch($this->lineItemCacheIdGenerator->generate($lineItemId)))
             );
-            self::assertEquals((int)$isActive, $cache['is_active_3']);
+            self::assertEquals($expectedLineItemStatus, $cache['status_3']);
         }
     }
 }