+++ tests/Unit/Entity/AssignmentTest.php
@@ -25,63 +25,77 @@ namespace OAT\SimpleRoster\Tests\Unit\Entity;
 use OAT\SimpleRoster\Entity\Assignment;
 use OAT\SimpleRoster\Entity\LineItem;
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Uid\UuidV6;
 
 class AssignmentTest extends TestCase
 {
-    /** @var LineItem */
-    private $lineItem;
-
-    /** @var Assignment */
-    private $subject;
-
-    protected function setUp(): void
-    {
-        parent::setUp();
-
-        $this->lineItem = new LineItem();
-        $this->subject = new Assignment();
-        $this->subject->setLineItem($this->lineItem);
-    }
-
     public function testItUpdatesStateAsCompletedIfMaxAttemptsIsReached(): void
     {
-        $this->lineItem->setMaxAttempts(1);
+        $lineItem = new LineItem(
+            new UuidV6('00000001-0000-6000-0000-000000000000'),
+            'testLabel',
+            'testUri',
+            'testSlug',
+            LineItem::STATUS_ENABLED,
+            1
+        );
 
-        $this->subject->incrementAttemptsCount();
+        $subject = new Assignment();
+        $subject->setLineItem($lineItem);
 
-        $this->subject->complete();
+        $subject->incrementAttemptsCount();
+
+        $subject->complete();
 
         self::assertSame(
             Assignment::STATE_COMPLETED,
-            $this->subject->getState()
+            $subject->getState()
         );
     }
 
     public function testItUpdatesStateAsReadyIfMaxAttemptsIs0(): void
     {
-        $this->lineItem->setMaxAttempts(0);
+        $lineItem = new LineItem(
+            new UuidV6('00000001-0000-6000-0000-000000000000'),
+            'testLabel',
+            'testUri',
+            'testSlug',
+            LineItem::STATUS_ENABLED
+        );
 
-        $this->subject->incrementAttemptsCount();
+        $subject = new Assignment();
+        $subject->setLineItem($lineItem);
 
-        $this->subject->complete();
+        $subject->incrementAttemptsCount();
+
+        $subject->complete();
 
         self::assertSame(
             Assignment::STATE_READY,
-            $this->subject->getState()
+            $subject->getState()
         );
     }
 
     public function testItUpdatesStateAsReadyIfMaxAttemptsIsNotReached(): void
     {
-        $this->lineItem->setMaxAttempts(2);
+        $lineItem = new LineItem(
+            new UuidV6('00000001-0000-6000-0000-000000000000'),
+            'testLabel',
+            'testUri',
+            'testSlug',
+            LineItem::STATUS_ENABLED,
+            2
+        );
+        $subject = new Assignment();
+        $subject->setLineItem($lineItem);
 
-        $this->subject->incrementAttemptsCount();
+        $subject->incrementAttemptsCount();
 
-        $this->subject->complete();
+        $subject->complete();
 
         self::assertSame(
             Assignment::STATE_READY,
-            $this->subject->getState()
+            $subject->getState()
         );
     }
 }