+++ tests/Unit/Webhook/Service/UpdateLineItemsServiceTest.php
@@ -92,17 +92,18 @@ class UpdateLineItemsServiceTest extends TestCase
             ->method('findBy')
             ->with(
                 [
-                    'slug' => ['qti-interactions-delivery']
+                    'slug' => ['qti-interactions-delivery'],
                 ]
             )
-            ->willReturn(
-                [
-                    (new LineItem())
-                        ->setId(new UuidV6('00000001-0000-6000-0000-000000000000'))
-                        ->setSlug('qti-interactions-delivery')
-                        ->setUri('http://lineitemuri.com')
-                ]
-            );
+            ->willReturn([
+                new LineItem(
+                    new UuidV6('00000001-0000-6000-0000-000000000000'),
+                    'testLabel',
+                    'http://lineitemuri.com',
+                    'qti-interactions-delivery',
+                    LineItem::STATUS_ENABLED
+                ),
+            ]);
 
         $this->entityManager->expects(self::once())
             ->method('persist');
@@ -113,7 +114,7 @@ class UpdateLineItemsServiceTest extends TestCase
                 'The line item id 00000001-0000-6000-0000-000000000000 was updated',
                 [
                     'oldUri' => 'http://lineitemuri.com',
-                    'newUri' => 'https://tao.instance/ontologies/tao.rdf#i5fb5'
+                    'newUri' => 'https://tao.instance/ontologies/tao.rdf#i5fb5',
                 ]
             );
 
@@ -140,20 +141,24 @@ class UpdateLineItemsServiceTest extends TestCase
 
     public function testItIgnoresDuplicatedUpdates(): void
     {
-        $lineItem = (new LineItem())
-            ->setId(new UuidV6('00000001-0000-6000-0000-000000000000'))
-            ->setSlug('qti-interactions-delivery');
+        $lineItem = new LineItem(
+            new UuidV6('00000001-0000-6000-0000-000000000000'),
+            'testLabel',
+            'testUri',
+            'qti-interactions-delivery',
+            LineItem::STATUS_ENABLED
+        );
 
         $this->lineItemRepository->expects(self::once())
             ->method('findBy')
             ->with(
                 [
-                    'slug' => ['qti-interactions-delivery', 'qti-interactions-delivery']
+                    'slug' => ['qti-interactions-delivery', 'qti-interactions-delivery'],
                 ]
             )
             ->willReturn(
                 [
-                    $lineItem
+                    $lineItem,
                 ]
             );
 
@@ -169,7 +174,7 @@ class UpdateLineItemsServiceTest extends TestCase
                 'The line item id 00000001-0000-6000-0000-000000000000 was updated',
                 [
                     'oldUri' => 'https://tao.instance/ontologies/tao.rdf#i5fb5',
-                    'newUri' => 'https://tao.instance/ontologies/tao.rdf#i5fb5'
+                    'newUri' => 'https://tao.instance/ontologies/tao.rdf#i5fb5',
                 ]
             );
 
@@ -210,7 +215,7 @@ class UpdateLineItemsServiceTest extends TestCase
             ->method('findBy')
             ->with(
                 [
-                    'slug' => ['qti-interactions-delivery']
+                    'slug' => ['qti-interactions-delivery'],
                 ]
             )
             ->willReturn([]);
@@ -222,7 +227,7 @@ class UpdateLineItemsServiceTest extends TestCase
             ->with(
                 'Impossible to update the line item. The slug qti-interactions-delivery does not exist.',
                 [
-                    'updateId' => '52a3de8dd0f270fd193f9f4bff05232f'
+                    'updateId' => '52a3de8dd0f270fd193f9f4bff05232f',
                 ]
             );